# Speed up --gil by skipping work for non-GIL threads